+++ src/main/java/com/dnf/game/Traverse.java
@@ -26,17 +26,6 @@ public class Traverse extends Base {
     @Resource
     private IniUtils iniUtils;
 
-    public MapTraversalType getMapData() {
-        // 地图遍历数据
-        MapTraversalType data = new MapTraversalType();
-        data.rwAddr = gamecall.personPtr();
-        data.mapData = memory.readLong(memory.readLong(data.rwAddr + Address.DtPyAddr) + 16);
-        data.start = memory.readLong(data.mapData + Address.DtKs2);
-        data.end = memory.readLong(data.mapData + Address.DtJs2);
-        data.objNum = (data.end - data.start) / 24;
-        return data;
-    }
-
     protected String[] getItemConfig() {
         String itemStr = iniUtils.read("自动配置", "过滤物品", String.class);
         return itemStr.split(",");
@@ -53,7 +42,7 @@ public class Traverse extends Base {
         String[] itemArr = this.getItemConfig();
 
         // 地图遍历数据
-        MapTraversalType data = getMapData();
+        MapTraversalType data = mapData.getMapData();
         for (data.objTmp = 1; data.objTmp < data.objNum; data.objTmp++) {
             data.objPtr = mapData.getTraversalPtr(data.start, data.objTmp, 2);
             data.objTypeA = memory.readInt(data.objPtr + Address.LxPyAddr);
@@ -133,7 +122,7 @@ public class Traverse extends Base {
         int harm = iniUtils.read("自动配置", "技能伤害", Integer.class);
         int size = iniUtils.read("自动配置", "技能大小", Integer.class);
 
-        MapTraversalType data = getMapData();
+        MapTraversalType data = mapData.getMapData();
         for (data.objTmp = 1; data.objTmp < data.objNum; data.objTmp++) {
             data.objPtr = mapData.getTraversalPtr(data.start, data.objTmp, 2);
             if (data.objPtr > 0) {
@@ -186,7 +175,7 @@ public class Traverse extends Base {
         }
 
         String[] itemArr = this.getItemConfig();
-        MapTraversalType data = getMapData();
+        MapTraversalType data = mapData.getMapData();
         for (data.objTmp = 1; data.objTmp < data.objNum; data.objTmp++) {
             data.objPtr = mapData.getTraversalPtr(data.start, data.objTmp, 2);
             data.objTypeA = memory.readInt(data.objPtr + Address.LxPyAddr);
